6-9  23:45 购物车、下多个单，管理地址 能正常用了

+++ 项目/online_bookStore/src/main/java/com/qdu/dao/impl/AddressDaoImpl.java
@@ -14,14 +14,15 @@ import java.util.List;
 
 public class AddressDaoImpl implements AddressDao {
     @Override
-    public List<Address> findAddressListById(String bid) {
+    public List<Address> findAddressListById(int uid) {
         Connection con = null;
         PreparedStatement ps = null;
         ResultSet rs = null;
         List<Address> address_List=new ArrayList();
         try {
             con = DatabaseUtil.getConnection();
-            ps = con.prepareStatement("select * from address");
+            ps = con.prepareStatement("select * from address where uid = ?");
+            ps.setInt(1,uid);
             rs = ps.executeQuery();
 
             while (rs.next()) {
@@ -72,15 +73,15 @@ public class AddressDaoImpl implements AddressDao {
         return rowsAffected;
     }
 
-    public int removeAddress(int userAddressId) {
+    public int removeAddress(int uid,int userAddressId) {
         Connection con = null;
         PreparedStatement ps = null;
         int rowsAffected = 0;
         try {
             con = DatabaseUtil.getConnection();
-            ps = con.prepareStatement("DELETE FROM address WHERE user_address_id = ?");
-            ps.setInt(1, userAddressId);
-
+            ps = con.prepareStatement("DELETE FROM address WHERE uid = ? and user_address_id = ?");
+            ps.setInt(1, uid);
+            ps.setInt(2,userAddressId);
             rowsAffected = ps.executeUpdate();
         } catch (SQLException e) {
             e.printStackTrace();
@@ -96,15 +97,21 @@ public class AddressDaoImpl implements AddressDao {
         int rowsAffected = 0;
         try {
             con = DatabaseUtil.getConnection();
-            ps = con.prepareStatement("INSERT INTO address (uid, province, city, area, info, postal_code, receiver_name, tel,user_address_id) VALUES (?, ?, ?, ?, ?, ?, ?, ?,?)");
-            ps.setInt(1, address.getUid());
-            ps.setString(2, address.getProvince());
-            ps.setString(3, address.getCity());
-            ps.setString(4, address.getArea());
-            ps.setString(5, address.getInfo());
-            ps.setString(6, address.getPostal_code());
-            ps.setString(7, address.getReceiver_name());
-            ps.setString(8, address.getTel());
+            // 假设你有一个唯一标识来更新特定的地址记录，比如id
+            // 这里我们使用address_id作为示例
+            String sql = "UPDATE address SET province = ?, city = ?, area = ?, info = ?, postal_code = ?, receiver_name = ?, tel = ? WHERE uid = ? and user_address_id = ? "; // 注意：这里假设address表中有一个名为address_id的字段作为主键
+
+            ps = con.prepareStatement(sql);
+
+            // 设置要更新的值
+            ps.setString(1, address.getProvince());
+            ps.setString(2, address.getCity());
+            ps.setString(3, address.getArea());
+            ps.setString(4, address.getInfo());
+            ps.setString(5, address.getPostal_code());
+            ps.setString(6, address.getReceiver_name());
+            ps.setString(7, address.getTel());
+            ps.setInt(8, address.getUid());
             ps.setInt(9,address.getUser_address_id());
 
             rowsAffected = ps.executeUpdate();
